+++ src/dswx_sar/common/_generate_log.py
